entitée professeur controlleur et service et html css

+++ src/main/java/com/example/ProfesseursClasse/controller/ProfController.java
@@ -70,7 +70,7 @@ public class ProfController {
   //effacer un professeur
 
     @GetMapping("/{id}/supprimer")
-    public String supprimerProf(@PathVariable Long id) {
+    public String supprimerProf(@PathVariable int id) {
         profService.supprimerProfById(id);
         return "redirect:/professeurs";
     }
